GOOGLE/business-opportunities-finder/backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# HTTP clients
httpx==0.26.0
aiohttp==3.9.1

# Web scraping
beautifulsoup4==4.12.3
lxml==5.1.0

# Reddit API
praw==7.7.1

# Google Gemini
google-generativeai==0.3.2

# Scheduling
apscheduler==3.10.4

# Rate limiting and retry
tenacity==8.2.3
ratelimit==2.2.1

# Email
aiosmtplib==3.0.1
email-validator==2.1.0

# Data processing
python-dateutil==2.8.2
pytz==2023.3

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Code quality
black==23.12.1
flake8==7.0.0
mypy==1.8.0
